[EJIT] Address review: pin phase-6 necessity, drop stray cast, renumber pipeline doc
- FinalGlobalDceDropsExpectGuardFreedHelper now pins the phase-6 premise:
  runModuleCleanup alone must keep the guard-held helper alive (1g), then
  the full pipeline must drop it (6). Guards fold in phase 2, so 1g's
  GlobalDCE cannot see the call as dead.
- Drop redundant static_cast in the phase-5 vectorization gate (enums
  already compare directly).
- runPipeline doc: renumber to 1a-1g / 2-4 / 5 / 6 to match the cpp.

+++ llvm/lib/ExecutionEngine/EJIT/EJitOptimizer.cpp
@@ -371,8 +371,7 @@ void EJitOptimizer::runOptimizationPipeline(Module &M,
   // Phase 5: vectorization + partial unrolling, L2+. Runs after the final
   // StructFieldPass so the vectorizers see the fully-specialized loops. L1
   // skips vectorization entirely (matching clang -O1).
-  if (static_cast<int>(level) >=
-      static_cast<int>(ejit::OptimizationLevel::L2)) {
+  if (level >= ejit::OptimizationLevel::L2) {
     runVectorization(M, level);
     EJIT_DIAG_DEBUG("pipeline phase5 done: vectorization opt=%d",
                     static_cast<int>(level));
